WIP Shopify

+++ app.js
@@ -3,15 +3,22 @@ const fs = require('fs');
 const path = require('path');
 const cheerio = require('cheerio');
 const timesToDownload = require('./timesToDownload.js')
-const downloadDirectory = 'E:/caiom/Documents/sacamisa/catalogo';
+const downloadDirectory = 'D:/projetos/sacamisa/catalogo';
 
-async function getImagensFromHuaze(url, modelo, time, liga){
+async function getImagensFromHuaze(url, modelo, time, liga) {
+  const dir = `${downloadDirectory}/${liga}/${time}/${modelo}`;
+  
+  // Check if the directory exists
+  if (fs.existsSync(dir)) {
+    console.log(`Diretório '${dir}' já existe. Pulando o download.`);
+    return; // Exit the function if the directory already exists
+  }
+  
   try {
     const responseHuaze = await axios.get(url);
     const $ = cheerio.load(responseHuaze.data);
-    // Check if the response status code is 200 (OK)
+    
     if (responseHuaze.status === 200) {
-      // Load the HTML content into Cheerio
       let imageUrls = [];
       fillImageUrls($, imageUrls);
       await salvaImagens(imageUrls, time, liga, modelo, url);
@@ -29,18 +36,20 @@ async function salvaImagens(imageUrls, time, liga, modelo, url) {
     const imageResponse = await axios.get('https://photo.yupoo.com/kakahuaze123' + imageSrc, { responseType: 'arraybuffer', headers: { Referer: url } });
     const imageBuffer = imageResponse.data;
 
-    // Specify the path where you want to save the image
-    const dir = `${downloadDirectory}/${liga}/${time}/${modelo}`
+    const dir = `${downloadDirectory}/${liga}/${time}/${modelo}`;
     const imagePath = path.join(dir, `${modelo}_${i + 1}.png`);
 
-    // Save the image to the specified path
-    if(!fs.existsSync(dir)){
-      fs.mkdirSync(dir, {recursive: true})
-      console.log(`Diretório Criado com Sucesso '${dir}'`)
+    if (!fs.existsSync(dir)) {
+      fs.mkdirSync(dir, { recursive: true });
+      console.log(`Diretório Criado com Sucesso '${dir}'`);
     }
-    fs.writeFileSync(imagePath, Buffer.from(imageBuffer, 'binary'));
 
-    console.log(`Image ${i + 1} downloaded and saved to ${imagePath}`);
+    if (!fs.existsSync(imagePath)) {
+      fs.writeFileSync(imagePath, Buffer.from(imageBuffer, 'binary'));
+      console.log(`Arquivo salvo com sucesso em '${imagePath}'`);
+    } else {
+      console.log(`Arquivo '${imagePath}' já existe. Não foi salvo.`);
+    }
   }
 }
 
@@ -49,6 +58,7 @@ function fillImageUrls($, imageUrls) {
     imageUrls.push($(this).attr('data-origin-src').split('123')[1]);
   });
 }
+
 async function main() {
   for (const liga in timesToDownload) {
     for (const time in timesToDownload[liga]) {
@@ -59,5 +69,5 @@ async function main() {
     }
   }
 }
-main();
 
+main();
